# PR: [libc++] Speed-up input_range based operations in vector<bool> 