As a follow-up to #120134 (which improved the performance of `forward_range` based operations in `vector<bool>`), this PR enhances the performance of `input_iterator`-pair or `input_range` based constructors and assignment functions. Instead of processing individual bits as in the previous implementation, the new implementation first collects bits into full storage words and then process entire words. This leads to performance improvements of up to 1.9x for the input_{iterator, range} constructors and up to 5.5x for the input_{iterator, range} assignment functions.

- input_iterator-pair ctors `vector( InputIt first, InputIt last, const Allocator& alloc)`: **1.9x**
- input_range-ctors `vector(std::from_range_t, R&& rg, const Allocator& alloc)`: **1.9x**
- input_iterator-pair assign function `assign(InputIt first, InputIt last)`: **5.1x**
- input_range assign function `assign_range(R&& rg)`: **5.5x**


#### Before:
```
--------------------------------------------------------------------------------------------------
Benchmark                                                        Time             CPU   Iterations
--------------------------------------------------------------------------------------------------
BM_ConstructInputIterIter<std::vector<bool>>/vb/514048     4726350 ns      4433882 ns          160
BM_ConstructFromInputRange<std::vector<bool>>/vb/514048    4824221 ns      4453109 ns          161
BM_AssignInputIterIter<std::vector<bool>>/vb/514048        2509647 ns      2316606 ns          302
BM_AssignInputRange<std::vector<bool>>/vb/514048           2604604 ns      2445725 ns          257
```

#### After:
```
--------------------------------------------------------------------------------------------------
Benchmark                                                        Time             CPU   Iterations
--------------------------------------------------------------------------------------------------
BM_ConstructInputIterIter<std::vector<bool>>/vb/514048     2533024 ns      2522681 ns          281
BM_ConstructFromInputRange<std::vector<bool>>/vb/514048    2551838 ns      2551748 ns          269
BM_AssignInputIterIter<std::vector<bool>>/vb/514048         490259 ns       485889 ns         1177
BM_AssignInputRange<std::vector<bool>>/vb/514048            475265 ns       469883 ns         1514
```